refactor: Move app.dart into app dir
